Move MatchPosition to a separated file
